Enhance authentication module by integrating ClerkAuthGuard and RolesGuard. Remove deprecated ClerkAuthGuard implementation and update exports in ClerkModule and AuthModule for improved security handling.
