+++ src/modules/auth/auth.module.ts
@@ -2,13 +2,22 @@ import { Module } from '@nestjs/common';
 import { UsersModule } from '../users/users.module';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
+import { RolesGuard } from './guards/roles.guard';
+import { ClerkModule } from '../clerk/clerk.module';
 
 @Module({
   imports: [
     UsersModule,
+    ClerkModule,
   ],
   controllers: [AuthController],
-  providers: [AuthService],
-  exports: [AuthService],
+  providers: [
+    AuthService,
+    RolesGuard,
+  ],
+  exports: [
+    AuthService,
+    RolesGuard,
+  ],
 })
 export class AuthModule {}
