+++ src/modules/clerk/guards/clerk-auth.guard.ts
@@ -1,0 +1,40 @@
+import {
+  Injectable,
+  CanActivate,
+  ExecutionContext,
+  UnauthorizedException,
+} from '@nestjs/common';
+import { ClerkSessionService } from '../clerk.session.service';
+
+@Injectable()
+export class ClerkAuthGuard implements CanActivate {
+  constructor(
+    private readonly clerkSessionService: ClerkSessionService,
+  ) {}
+
+  async canActivate(context: ExecutionContext): Promise<boolean> {
+    const request = context.switchToHttp().getRequest();
+    
+    try {
+      // Extract token from Authorization header
+      const authHeader = request.headers.authorization;
+      if (!authHeader || !authHeader.startsWith('Bearer ')) {
+        throw new UnauthorizedException('Missing or invalid authorization header');
+      }
+
+      const token = authHeader.substring(7); // Remove 'Bearer ' prefix
+
+      // Use ClerkSessionService to verify token and get authentication data
+      const authData = await this.clerkSessionService.verifyTokenAndGetAuthData(token);
+
+      // Attach authentication data to request object
+      request.user = authData.user;
+      request.session = authData.session;
+      request.sessionClaims = authData.sessionClaims;
+
+      return true;
+    } catch (error) {
+      throw new UnauthorizedException(`Authentication failed: ${error.message}`);
+    }
+  }
+}
